fix: allow decimal needsPct/wantsPct/savingsPct
The mobile app's numpad lets a customer type an exact VND amount for a
50-30-20 bucket (e.g. 15.300.000đ on a 100.000.000đ income = 15.3%).
Storing only the nearest whole percent silently discarded that precision
on every save. needs_pct/wants_pct/savings_pct move from integer to
numeric(5,2) on both customers and income_allocation_settings (new
migration V0007), matching the monthly_income numeric(15,2) convention
already used on both tables.

+++ src/FinViet.Api/Controllers/ProfileController.cs
@@ -150,15 +150,15 @@ public record UpdateProfileRequest(
     Gender? Gender = null,
     DateOnly? DateOfBirth = null,
     bool? OnboardingDone = null,
-    int? NeedsPct = null,
-    int? WantsPct = null,
-    int? SavingsPct = null);
+    decimal? NeedsPct = null,
+    decimal? WantsPct = null,
+    decimal? SavingsPct = null);
 
 public record ScheduleIncomeAllocationRequest(
     decimal MonthlyIncome,
-    int NeedsPct,
-    int WantsPct,
-    int SavingsPct);
+    decimal NeedsPct,
+    decimal WantsPct,
+    decimal SavingsPct);
 
 public record UpdateProfileSettingsRequest(
     AppTheme? Theme = null,
